config.js
configuration has been changed to allow for multiple WebAPI services to be defined.
HERMES will use the first service listed in the configuration array as the default.
a new dropdown button has been added next to the search box that allows the user to choose which service to use for searches
defaulting to public means no local WebAPI installation is required for basic HERMES use

invalid concepts
clear visual indicators have been integrated to draw attention to concepts that are no longer valid both on the search screen as well as the concept display. In the guided search the name of an invalid concept will have a red font color. On the concept screen a red Invalid badge appears below the concept name.  In the related concept table the link for the concept will have a red font color.

misc
a loading indicator has been added while waiting for concept and related concept information to be returned
stylesheets for bootstrap and dataTable overrides separated from main application styles